Some re-organizing
