+++ src/Configuration.cpp
@@ -9,6 +9,7 @@
  */
 
 #include "Configuration.h"
+
 #include <algorithm>
 #include <cstdlib>
 #include <limits.h>
