+++ src/FtlClient.cpp
@@ -7,7 +7,7 @@
 
 #include "FtlClient.h"
 
-#include "Util.h"
+#include "Utilities/Util.h"
 
 #include <fmt/core.h>
 #include <netdb.h>
