+++ src/FtlStreamStore.cpp
@@ -9,6 +9,7 @@
 
 #include "FtlStreamStore.h"
 #include "FtlStream.h"
+
 extern "C"
 {
     #include <debug.h>
